requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.0,<24.0

# Database
psycopg2-binary>=2.9,<3.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Security
python-dotenv>=1.0,<2.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Code Quality
black>=23.0,<24.0
flake8>=6.1,<7.0
isort>=5.12,<6.0

# Utilities
uuid>=1.30

# Production
gunicorn>=21.0,<22.0
